+++ cypress/cucumber-json/App.cucumber.json
@@ -23,7 +23,7 @@
             "name": "\"アプリケーション\" ページにアクセスする",
             "result": {
               "status": "passed",
-              "duration": 4101000000
+              "duration": 4116000000
             }
           },
           {
@@ -33,7 +33,7 @@
             "name": "\"やること\" を追加する",
             "result": {
               "status": "passed",
-              "duration": 258000000
+              "duration": 301000000
             }
           },
           {
@@ -43,7 +43,7 @@
             "name": "\"やること\" が表示される",
             "result": {
               "status": "passed",
-              "duration": 21000000
+              "duration": 33000000
             }
           },
           {
@@ -53,7 +53,7 @@
             "name": "ステータスが \"未着手\" になる",
             "result": {
               "status": "passed",
-              "duration": 19000000
+              "duration": 25000000
             }
           },
           {
@@ -63,7 +63,7 @@
             "name": "アイテム数が \"1\" 件になる",
             "result": {
               "status": "passed",
-              "duration": 21000000
+              "duration": 47000000
             }
           }
         ]
@@ -83,7 +83,7 @@
             "name": "\"アプリケーション\" ページにアクセスする",
             "result": {
               "status": "passed",
-              "duration": 2176000000
+              "duration": 2359000000
             }
           },
           {
@@ -93,7 +93,7 @@
             "name": "期限を設定する",
             "result": {
               "status": "passed",
-              "duration": 325000000
+              "duration": 426000000
             }
           },
           {
@@ -103,7 +103,7 @@
             "name": "期限が設定される",
             "result": {
               "status": "passed",
-              "duration": 16000000
+              "duration": 18000000
             }
           },
           {
@@ -113,7 +113,7 @@
             "name": "ステータスが \"着手\" になる",
             "result": {
               "status": "passed",
-              "duration": 22000000
+              "duration": 19000000
             }
           },
           {
@@ -123,7 +123,7 @@
             "name": "アイテム数が \"1\" 件になる",
             "result": {
               "status": "passed",
-              "duration": 14000000
+              "duration": 31000000
             }
           }
         ]
@@ -143,7 +143,7 @@
             "name": "\"アプリケーション\" ページにアクセスする",
             "result": {
               "status": "passed",
-              "duration": 2177000000
+              "duration": 2212000000
             }
           },
           {
@@ -153,7 +153,7 @@
             "name": "\"やること\" を完了する",
             "result": {
               "status": "passed",
-              "duration": 113000000
+              "duration": 123000000
             }
           },
           {
@@ -163,7 +163,7 @@
             "name": "\"やること\" が更新される",
             "result": {
               "status": "passed",
-              "duration": 19000000
+              "duration": 17000000
             }
           },
           {
@@ -173,7 +173,7 @@
             "name": "ステータスが \"完了\" になる",
             "result": {
               "status": "passed",
-              "duration": 16000000
+              "duration": 15000000
             }
           },
           {
@@ -183,7 +183,7 @@
             "name": "アイテム数が \"1\" 件になる",
             "result": {
               "status": "passed",
-              "duration": 15000000
+              "duration": 16000000
             }
           }
         ]
@@ -203,7 +203,7 @@
             "name": "\"アプリケーション\" ページにアクセスする",
             "result": {
               "status": "passed",
-              "duration": 2163000000
+              "duration": 2177000000
             }
           },
           {
@@ -213,7 +213,7 @@
             "name": "\"やること\" を削除する",
             "result": {
               "status": "passed",
-              "duration": 109000000
+              "duration": 108000000
             }
           },
           {
@@ -223,7 +223,7 @@
             "name": "\"やること\" が削除される",
             "result": {
               "status": "passed",
-              "duration": 18000000
+              "duration": 20000000
             }
           },
           {
