+++ index.js
@@ -1,6 +1,6 @@
 import "./src/style.css";
 import { App } from "./src/App.js";
-import { DB } from "./src/DB.js";
+import { DB } from "./src/infrastructure/DB.js";
 import { setUp } from "./src/Dev.js";
 
 const db = new DB("todo");
